refactor(dual-query-planner): moved semantic diff out into its own module
